Split malware scanner and attachments for improved readbility
