+++ srv/attachments/azure-blob-storage.js
@@ -1,38 +1,38 @@
-const { Storage } = require("@google-cloud/storage")
+const { BlobServiceClient } = require("@azure/storage-blob")
+const { AbortController } = require("abort-controller")
 const cds = require("@sap/cds")
 const LOG = cds.log("attachments")
-const utils = require("../lib/helper")
+const utils = require("../../lib/helper")
 const {
   MAX_FILE_SIZE,
   sizeInBytes,
   createSizeCheckHandler,
-} = require("../lib/helper")
+} = require("../../lib/helper")
 
-module.exports = class GoogleAttachmentsService extends (
+module.exports = class AzureAttachmentsService extends (
   require("./object-store")
 ) {
   /**
-   * Creates or retrieves a cached Google Cloud Platform client for the given tenant
-   * @returns {Promise<{bucket: import('@google-cloud/storage').Bucket}>}
+   * Creates or retrieves a cached Azure Blob Storage client for the given tenant
+   * @returns {Promise<{blobServiceClient: import('@azure/storage-blob').BlobServiceClient, containerClient: import('@azure/storage-blob').ContainerClient}>}
    */
   async retrieveClient() {
     const tenantID = this.separateObjectStore ? cds.context.tenant : "shared"
-    LOG.debug("Retrieving tenant-specific Google Cloud Platform client", {
+    LOG.debug("Retrieving tenant-specific Azure Blob Storage client", {
       tenantID,
     })
+
     const existingClient = this.clientsCache.get(tenantID)
     if (existingClient) {
-      LOG.debug("Using cached GCP client", {
+      LOG.debug("Using cached Azure Blob Storage client", {
         tenantID,
-        bucketName: existingClient.bucket.name,
+        containerName: existingClient.containerClient.containerName,
       })
       return existingClient
     }
 
     try {
-      LOG.debug(
-        `Fetching object store credentials for tenant ${tenantID}. Using ${this.separateObjectStore ? "shared" : "tenant-specific"} object store.`,
-      )
+      LOG.debug("Fetching object store credentials for tenant", { tenantID })
       const credentials = this.separateObjectStore
         ? (await utils.getObjectStoreCredentials(tenantID))?.credentials
         : cds.env.requires?.objectStore?.credentials
@@ -41,12 +41,7 @@ module.exports = class GoogleAttachmentsService extends (
         throw new Error("SAP Object Store instance is not bound.")
       }
 
-      // Validate required credentials
-      const requiredFields = [
-        "bucket",
-        "projectId",
-        "base64EncodedPrivateKeyData",
-      ]
+      const requiredFields = ["container_name", "container_uri", "sas_token"]
       const missingFields = requiredFields.filter(
         (field) => !credentials[field],
       )
@@ -54,53 +49,65 @@ module.exports = class GoogleAttachmentsService extends (
       if (missingFields.length > 0) {
         if (credentials.access_key_id) {
           throw new Error(
-            "AWS S3 credentials found where Google Cloud Platform credentials expected, please check your service bindings.",
+            "AWS S3 credentials found where Azure Blob Storage credentials expected, please check your service bindings.",
           )
-        } else if (credentials.container_name) {
+        } else if (credentials.projectId) {
           throw new Error(
-            "Azure credentials found where Google Cloud Platform credentials expected, please check your service bindings.",
+            "Google Cloud Platform credentials found where Azure Blob Storage credentials expected, please check your service bindings.",
           )
         }
         throw new Error(
-          `Missing Google Cloud Platform credentials: ${missingFields.join(", ")}`,
+          `Missing Azure Blob Storage credentials: ${missingFields.join(", ")}`,
         )
       }
 
-      LOG.debug("Creating Google Cloud Platform client for tenant", {
+      LOG.debug("Creating Azure Blob Storage client for tenant", {
         tenantID,
-        bucketName: credentials.bucket,
+        containerName: credentials.container_name,
       })
 
-      const storageClient = new Storage({
-        projectId: credentials.projectId,
-        credentials: JSON.parse(
-          Buffer.from(
-            credentials.base64EncodedPrivateKeyData,
-            "base64",
-          ).toString("utf8"),
-        ),
-      })
+      const blobServiceClient = new BlobServiceClient(
+        credentials.container_uri + "?" + credentials.sas_token,
+      )
+      const containerClient = blobServiceClient.getContainerClient(
+        credentials.container_name,
+      )
 
-      const newGoogleClient = {
-        bucket: storageClient.bucket(credentials.bucket),
+      const newAzureCredentials = {
+        containerClient,
       }
 
-      this.clientsCache.set(tenantID, newGoogleClient)
+      this.clientsCache.set(tenantID, newAzureCredentials)
 
-      LOG.debug("Google Cloud Platform client has been created successful", {
+      LOG.debug("Azure Blob Storage client has been created successful", {
         tenantID,
-        bucketName: newGoogleClient.bucket.name,
+        containerName: containerClient.containerName,
       })
-
-      return newGoogleClient
+      return newAzureCredentials
     } catch (error) {
       LOG.error(
-        "Failed to create tenant-specific Google Cloud Platform client",
+        "Failed to create tenant-specific Azure Blob Storage client",
         error,
-        "Check Service Manager and Google Cloud Platform instance configuration",
+        "Check Service Manager and Azure Blob Storage instance configuration",
         { tenantID },
       )
       throw error
+    }
+  }
+
+  async exists(blobName) {
+    const { containerClient } = await this.retrieveClient()
+    const blobClient = containerClient.getBlockBlobClient(blobName)
+    try {
+      await blobClient.getProperties()
+      // If no error, blob exists
+      return true
+    } catch (err) {
+      // Anything besides 404 BlobNotFound is an actual error
+      if (err.statusCode !== 404 && err.code !== "BlobNotFound") {
+        throw err
+      }
+      return false
     }
   }
 
@@ -118,20 +125,18 @@ module.exports = class GoogleAttachmentsService extends (
 
     const startTime = Date.now()
 
-    LOG.debug("Starting file upload to Google Cloud Platform", {
+    LOG.debug("Starting file upload to Azure Blob Storage", {
       attachmentEntity: attachments.name,
       tenant: cds.context.tenant,
     })
-
-    const { bucket } = await this.retrieveClient()
-
+    const { containerClient } = await this.retrieveClient()
     try {
-      const { content, ...metadata } = data
+      let { content, ...metadata } = data
       const blobName = metadata.url
 
       if (!blobName) {
         LOG.error(
-          "File key/URL is required for Google Cloud Platform upload",
+          "File key/URL is required for Azure Blob Storage upload",
           null,
           "Ensure attachment data includes a valid URL/key",
           { metadata: { ...metadata, content: !!content } },
@@ -141,7 +146,7 @@ module.exports = class GoogleAttachmentsService extends (
 
       if (!content) {
         LOG.error(
-          "File content is required for Google Cloud Platform upload",
+          "File content is required for Azure Blob Storage upload",
           null,
           "Ensure attachment data includes file content",
           { key: blobName, hasContent: !!content },
@@ -149,10 +154,9 @@ module.exports = class GoogleAttachmentsService extends (
         throw new Error("File content is required for upload")
       }
 
-      const file = bucket.file(blobName)
+      const blobClient = containerClient.getBlockBlobClient(blobName)
 
-      const [exists] = await file.exists()
-      if (exists) {
+      if (await this.exists(blobName)) {
         const error = new Error("Attachment already exists")
         error.status = 409
         throw error
@@ -170,32 +174,24 @@ module.exports = class GoogleAttachmentsService extends (
             ) ?? MAX_FILE_SIZE)
           : MAX_FILE_SIZE
 
-      LOG.debug("Uploading file to Google Cloud Platform", {
-        bucketName: bucket.name,
+      LOG.debug("Uploading file to Azure Blob Storage", {
+        containerName: containerClient.containerName,
         blobName,
         maxFileSize,
       })
 
       const sizeLimit =
         attachments.elements.content["@Validation.Maximum"] || "400MB"
-      const writeStream = file.createWriteStream()
 
-      let resolveUpload
-      const uploadPromise = new Promise((resolve) => {
-        resolveUpload = resolve
-      })
-
+      const abortController = new AbortController()
       const { handler, getSizeExceeded, createError } = createSizeCheckHandler({
         maxFileSize,
         filename: attachmentRef?.filename,
         sizeLimit,
         onSizeExceeded: () => {
-          // Unpipe and destroy the write stream to abort the upload
-          content.unpipe(writeStream)
-          writeStream.destroy()
+          abortController.abort()
           // Resume content to drain it (prevents backpressure from hanging the connection)
           content.resume()
-          resolveUpload()
         },
       })
 
@@ -203,22 +199,9 @@ module.exports = class GoogleAttachmentsService extends (
 
       // The file upload has to be done first, so super.put can compute the hash and trigger malware scan
       try {
-        await Promise.race([
-          uploadPromise,
-          new Promise((resolve, reject) => {
-            content.pipe(writeStream)
-            writeStream.on("finish", resolve)
-            writeStream.on("error", (err) => {
-              // Ignore errors if size exceeded - we intentionally destroyed the stream
-              if (getSizeExceeded()) {
-                resolve()
-              } else {
-                reject(err)
-              }
-            })
-            content.on("error", reject)
-          }),
-        ])
+        await blobClient.uploadStream(content, undefined, undefined, {
+          abortSignal: abortController.signal,
+        })
       } catch (err) {
         if (getSizeExceeded()) {
           throw createError()
@@ -226,23 +209,12 @@ module.exports = class GoogleAttachmentsService extends (
         throw err
       }
 
-      // Check after promise resolves if size was exceeded
-      if (getSizeExceeded()) {
-        // Try to delete the partial upload
-        try {
-          await file.delete({ ignoreNotFound: true })
-        } catch {
-          // Ignore delete errors
-        }
-        throw createError()
-      }
-
       await super.put(attachments, metadata)
 
       const duration = Date.now() - startTime
-      LOG.debug("File upload to Google Cloud Platform completed successfully", {
+      LOG.debug("File upload to Azure Blob Storage completed successfully", {
         fileId: metadata.ID,
-        bucketName: bucket.name,
+        containerName: containerClient.containerName,
         blobName,
         duration,
       })
@@ -252,12 +224,12 @@ module.exports = class GoogleAttachmentsService extends (
       }
       const duration = Date.now() - startTime
       LOG.error(
-        "File upload to Google Cloud Platform failed",
+        "File upload to Azure Blob Storage failed",
         err,
-        "Check Google Cloud Platform connectivity, credentials, and container permissions",
+        "Check Azure Blob Storage connectivity, credentials, and container permissions",
         {
           fileId: data?.ID,
-          bucketName: bucket.name,
+          containerName: containerClient.containerName,
           blobName: data?.url,
           duration,
         },
@@ -271,12 +243,12 @@ module.exports = class GoogleAttachmentsService extends (
    */
   async get(attachments, keys) {
     const startTime = Date.now()
-    LOG.debug("Starting stream from Google Cloud Platform", {
+    LOG.debug("Starting stream from Azure Blob Storage", {
       attachmentEntity: attachments.name,
       keys,
       tenant: cds.context.tenant,
     })
-    const { bucket } = await this.retrieveClient()
+    const { containerClient } = await this.retrieveClient()
 
     try {
       LOG.debug("Fetching attachment metadata", { keys })
@@ -292,41 +264,38 @@ module.exports = class GoogleAttachmentsService extends (
         return null
       }
 
-      const blobName = response.url
-
-      LOG.debug("Streaming file from Google Cloud Platform", {
-        bucketName: bucket.name,
-        blobName,
+      LOG.debug("Streaming file from Azure Blob Storage", {
+        containerName: containerClient.containerName,
+        fileId: keys.ID,
+        blobName: response.url,
       })
 
-      const file = bucket.file(blobName)
-      const readStream = await file.createReadStream()
+      const blobClient = containerClient.getBlockBlobClient(response.url)
+      const downloadResponse = await blobClient.download()
 
       const duration = Date.now() - startTime
-      LOG.debug("File streamed from Google Cloud Platform successfully", {
+      LOG.debug("File streamed from Azure Blob Storage successfully", {
         fileId: keys.ID,
-        bucketName: bucket.name,
-        blobName,
         duration,
       })
 
-      return readStream
+      return downloadResponse.readableStreamBody
     } catch (error) {
       const duration = Date.now() - startTime
       const suggestion =
         error.code === "BlobNotFound"
-          ? "File may have been deleted from Google Cloud Platform or URL is incorrect"
+          ? "File may have been deleted from Azure Blob Storage or URL is incorrect"
           : error.code === "AuthenticationFailed"
-            ? "Check Google Cloud Platform credentials and SAS token"
-            : "Check Google Cloud Platform connectivity and configuration"
+            ? "Check Azure Blob Storage credentials and SAS token"
+            : "Check Azure Blob Storage connectivity and configuration"
 
       LOG.error(
-        "File download from Google Cloud Platform failed",
+        "File download from Azure Blob Storage failed",
         error,
         suggestion,
         {
           fileId: keys?.ID,
-          bucketName: bucket.name,
+          containerName: containerClient.containerName,
           attachmentName: attachments.name,
           duration,
         },
@@ -337,22 +306,23 @@ module.exports = class GoogleAttachmentsService extends (
   }
 
   /**
-   * Deletes a file from Google Cloud Platform
+   * Deletes a file from Azure Blob Storage
    * @param {string} Key - The key of the file to delete
    * @returns {Promise} - Promise resolving when deletion is complete
    */
   async delete(blobName) {
-    const { bucket } = await this.retrieveClient()
+    const { containerClient } = await this.retrieveClient()
     LOG.debug(
-      `[GCP] Executing delete for file ${blobName} in bucket ${bucket.name}`,
+      `[Azure] Executing delete for file ${blobName} in bucket ${containerClient.containerName}`,
     )
 
-    const file = bucket.file(blobName)
-    const response = await file.delete()
-    if (response[0]?.statusCode !== 204) {
-      LOG.warn("File has not been deleted from Google Cloud Storage", {
+    const blobClient = containerClient.getBlockBlobClient(blobName)
+    const response = await blobClient.delete()
+
+    if (response._response.status !== 202) {
+      LOG.warn("File has not been deleted from Azure Blob Storage", {
         blobName,
-        bucketName: bucket.name,
+        containerName: containerClient.containerName,
         response,
       })
     }
