+++ srv/attachments/standard.js
@@ -1,4 +1,4 @@
-const { getAttachmentKind } = require("../lib/helper")
+const { getAttachmentKind } = require("../../lib/helper")
 const LOG = require("@sap/cds").log("attachments")
 
 LOG.info(`Using ${getAttachmentKind()} for attachments management.`)
