+++ srv/malware-scanner/malwareScanner-mocked.js
