+++ srv/malware-scanner/malwareScanner.js
@@ -1,7 +1,6 @@
 const cds = require("@sap/cds")
 const crypto = require("crypto")
 const https = require("https")
-const axios = require("axios")
 const { URL } = require("url")
 const Semaphore = require("./semaphore")
 const LOG = cds.log("attachments")
@@ -100,6 +99,7 @@ class MalwareScanner extends cds.ApplicationService {
       LOG.debug(`Malware scan completed for ${target}, ${keys} - file is clean`)
     }
 
+    // Assign hash as another condition to ensure the correct file is marked as fine
     await this.updateStatus(_target, Object.assign({ hash }, keys), status)
   }
 
@@ -193,13 +193,24 @@ class MalwareScanner extends cds.ApplicationService {
       const url = new URL(`https://${this.credentials.uri}/scan`)
       const { headers, httpsAgent } = this._buildRequestConfig()
 
-      const response = await axios.post(url.toString(), file, {
+      const fetchOptions = {
+        method: "POST",
         headers,
-        httpsAgent,
-        maxBodyLength: Infinity,
-        maxContentLength: Infinity,
-        validateStatus: (status) => status >= 200 && status < 300,
-      })
+        body: file,
+        duplex: "half",
+      }
+      if (httpsAgent) fetchOptions.agent = httpsAgent
+
+      const response = await fetch(url.toString(), fetchOptions)
+
+      if (!response.ok) {
+        const error = new Error(`Malware scan request failed with status ${response.status}`)
+        error.response = {
+          status: response.status,
+          headers: Object.fromEntries(response.headers),
+        }
+        throw error
+      }
 
       /**
        * @typedef {Object} MalwareScanResponse
@@ -211,7 +222,7 @@ class MalwareScanner extends cds.ApplicationService {
        * @property {string} SHA256 - SHA-256 hash of the scanned file.
        */
       /** @type {MalwareScanResponse} */
-      const responseJson = response.data
+      const responseJson = await response.json()
       const scanDuration = Date.now() - scanStartTime
       LOG.debug(`Malware scan response`, {
         scanDuration,
