+++ tests/unit/malwareScanner.test.js
@@ -3,10 +3,8 @@ const cds = require("@sap/cds")
 const { join } = cds.utils.path
 cds.test(join(__dirname, "../incidents-app"))
 
-jest.mock("axios")
-
-const MalwareScanner = require("../../srv/malwareScanner")
-const Semaphore = require("../../srv/semaphore")
+const MalwareScanner = require("../../srv/malware-scanner/malwareScanner")
+const Semaphore = require("../../srv/malware-scanner/semaphore")
 
 let scanner
 let attachmentsSvc
